requirements.txt
langchain==0.0.146
python-dotenv==1.0.0
